testing: move account benchmarks to accounts_test.go
